+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - New Admin landing page
 - New Instructor Course Section overview page
+- Allow LMS and System admins to unlink LMS section
 
 ## 0.14.0 (2021-10-13)